Refactor: add const qualifiers for (de)compression
- Updated `compress_chunk` function to use `const` qualifiers for `chunk` and `chunk_model` parameters to ensure data integrity
- Updated `decompress_cmp_entiy` function to use `const` qualifiers for `ent` and `model_of_data` parameters to ensure data integrity
- Renamed `input_buf` to `src` and `icu_output_buf` to `dst` and `buffer_length` to `stream_size` for clarity

+++ test/fuzz/fuzz_round_trip.c
@@ -77,7 +77,7 @@ int LLVMFuzzerTestOneInput(const uint8_t *src, size_t size)
 	use_decmp_buf = FUZZ_dataProducer_int32Range(producer, 0, 1);
 	use_decmp_up_model = FUZZ_dataProducer_int32Range(producer, 0, 1);
 
-	return_value = chunk_round_trip(src, size, model, up_model, cmp_data,
+	return_value = chunk_round_trip(src, (uint32_t)size, model, up_model, cmp_data,
 				        cmp_data_capacity, cmp_par_ptr,
 				        use_decmp_buf, use_decmp_up_model);
 	switch (cmp_get_error_code(return_value)) {
